Move notification functions into separate file.
