+++ v3/src/components/case-table/attribute-menu/edit-attribute-properties-modal.tsx
@@ -1,12 +1,12 @@
 import { Button, FormControl, FormLabel, HStack, Input, ModalBody, ModalCloseButton, ModalFooter, ModalHeader,
   Radio, RadioGroup, Select, Textarea, Tooltip } from "@chakra-ui/react"
 import React, { useEffect, useState } from "react"
-import { AttributeType, attributeTypes } from "../../../models/data/attribute"
 import { useDataSetContext } from "../../../hooks/use-data-set-context"
+import { AttributeType, attributeTypes } from "../../../models/data/attribute"
+import { updateAttributesNotification } from "../../../models/data/data-set-notifications"
 import { uniqueName } from "../../../utilities/js-utils"
-import { CodapModal } from "../../codap-modal"
 import { t } from "../../../utilities/translation/translate"
-import { updateAttributesNotification } from "../../../models/data/data-set-utils"
+import { CodapModal } from "../../codap-modal"
 
 // for use in menus of attribute types
 const selectableAttributeTypes = ["none", ...attributeTypes] as const
