+++ v3/src/components/case-table/column-header-divider.tsx
@@ -3,7 +3,8 @@ import { createPortal } from "react-dom"
 import { IAttribute } from "../../models/data/attribute"
 import { isCollectionModel } from "../../models/data/collection"
 import { IMoveAttributeOptions } from "../../models/data/data-set-types"
-import { getCollectionAttrs, moveAttributeNotification } from "../../models/data/data-set-utils"
+import { moveAttributeNotification } from "../../models/data/data-set-notifications"
+import { getCollectionAttrs } from "../../models/data/data-set-utils"
 import { useCollectionContext } from "../../hooks/use-collection-context"
 import { useDataSetContext } from "../../hooks/use-data-set-context"
 import { getDragAttributeInfo, useTileDroppable } from "../../hooks/use-drag-drop"
