+++ v3/src/components/case-table/inspector-panel/ruler-menu-list.tsx
@@ -2,7 +2,7 @@ import { MenuItem, MenuList, useToast } from "@chakra-ui/react"
 import React from "react"
 import { useDataSetContext } from "../../../hooks/use-data-set-context"
 import { IAttribute } from "../../../models/data/attribute"
-import { createAttributesNotification } from "../../../models/data/data-set-utils"
+import { createAttributesNotification } from "../../../models/data/data-set-notifications"
 import { uniqueName } from "../../../utilities/js-utils"
 import { t } from "../../../utilities/translation/translate"
